Update README.md
# Sentiment-analysis

In this project, a predictive model is built using the amazon-review so that it can be tested on further reviews. The whole project is done through NLP sentiment analysis using spaCY package.

Data set: https://www.kaggle.com/datasets/snap/amazon-fine-food-reviews
The list of different accuracy is given below:

Train Accuracy 73.15%
Test Accuracy 59.54%
Sensitivity- 56.83%
Specificity- 86.53%


The difference between the Train and Test data is quite significant. But these kind of observations are accepted as the vocabulary of english is very large. So the texts in Train and Test data differs a lot.Self Project